Implement rollup bundling
This implements a fairly standard Rollup bundling implementation, with
two different outputs for CommonJS and for ESM.

This moves to using the Rollup TypeScript plugin, and adds a couple
useful plugins for resolving node-like and commonjs packages.

May need some tweaking in externalizing dependencies, and I haven't
verified bundle sizes, but Rollup's ESM output should get us "as good as
it gets" bundle sizes when configured properly.

Should be an improvement over tsc's commonjs output, which is what was configured prior.